+++ views/edit/modal_edit_file.php
@@ -1,6 +1,7 @@
 <?php
 
 
+use humhub\modules\content\widgets\richtext\RichTextField;
 use humhub\modules\topic\widgets\TopicPicker;
 use humhub\widgets\ModalButton;
 use humhub\widgets\ModalDialog;
@@ -20,7 +21,7 @@ use yii\bootstrap\ActiveForm;
 
         <div class="modal-body">
             <?= $form->field($file->baseFile, 'file_name'); ?>
-            <?= $form->field($file, 'description'); ?>
+            <?= $form->field($file, 'description')->widget(RichTextField::class); ?>
             <?= $form->field($file, 'topics')->widget(TopicPicker::class, ['contentContainer' => $file->content->container])->label(false); ?>
             <?= $form->field($file, 'visibility')->checkbox(['disabled' => $file->parentFolder->content->isPrivate()]) ?>
             <?= $form->field($file, 'download_count')->staticControl(['style' => 'display:inline']); ?>